chore(client): move deploy modal -> modals
