+++ client/src/app/modals/deploy-diagram/View.less
